requirements.txt
Django>=3.1
django-ninja
uvicorn
faker